Satisfierd with squeeze and begun with side spawn behaviuor
Squeeze effect is tuned for different siezes of obstacles. Some bugs but I will have to deal with them another time.

Just implemented an ImpulseForce on obstacles that spawn to the side of the screen. Didn't have time to adjust he "sideForceIndex" that controls the force as a function of height in the oManager. Until next time

The method for spawning obstacle on the side in the "SpawnerScript" is also on working properly. Until Next time.