All obstacles bounce and powerUps and hearts fall
Bouncy floor, all obstacles bounce on the bouncy floor as many times as according to their color. I don't know how to make squares and triangles move porperly with this physics based set up. I thik it would be better to cange the obstacles positions according to lerped position, disregarding. this way I could chain different lerped positions as when they finnish say a bounce is betwenn point a and b with an offset in height between them. As the lerp index reaches 1 meaning that the obstacle has reached its point the next lerp in position would begin. This ay I can change the positions and the offsets depending on shape.

the player is now kinematic to not end up jumping around when he is hit by obstacles.

PowerUps and hearts no longer have rigidbodies to not be affected by hre bouncy floor. They fall according to a simple linear lerp.